Fixing Host ID initialization (ensuring it is set when FastLogger is created)
