+++ src/WebJobs.Script/Host/ScriptHost.cs
@@ -278,7 +278,7 @@ namespace Microsoft.Azure.WebJobs.Script
 
                 if (string.IsNullOrEmpty(ScriptConfig.HostConfig.HostId))
                 {
-                    ScriptConfig.HostConfig.HostId = GetDefaultHostId(_settingsManager, ScriptConfig);
+                    ScriptConfig.HostConfig.HostId = Utility.GetDefaultHostId(_settingsManager, ScriptConfig);
                 }
                 if (string.IsNullOrEmpty(ScriptConfig.HostConfig.HostId))
                 {
@@ -1302,44 +1302,6 @@ namespace Microsoft.Azure.WebJobs.Script
                     }
                 }
             }
-        }
-
-        internal static string GetDefaultHostId(ScriptSettingsManager settingsManager, ScriptHostConfiguration scriptConfig)
-        {
-            // We're setting the default here on the newly created configuration
-            // If the user has explicitly set the HostID via host.json, it will overwrite
-            // what we set here
-            string hostId = null;
-            if (scriptConfig.IsSelfHost)
-            {
-                // When running locally, derive a stable host ID from machine name
-                // and root path. We use a hash rather than the path itself to ensure
-                // IDs differ (due to truncation) between folders that may share the same
-                // root path prefix.
-                // Note that such an ID won't work in distributed scenarios, so should
-                // only be used for local/CLI scenarios.
-                string sanitizedMachineName = Environment.MachineName
-                    .Where(char.IsLetterOrDigit)
-                    .Aggregate(new StringBuilder(), (b, c) => b.Append(c)).ToString();
-                hostId = $"{sanitizedMachineName}-{Math.Abs(scriptConfig.RootScriptPath.GetHashCode())}";
-            }
-            else if (!string.IsNullOrEmpty(settingsManager.AzureWebsiteUniqueSlotName))
-            {
-                // If running on Azure Web App, derive the host ID from unique site slot name
-                hostId = settingsManager.AzureWebsiteUniqueSlotName;
-            }
-
-            if (!string.IsNullOrEmpty(hostId))
-            {
-                if (hostId.Length > ScriptConstants.MaximumHostIdLength)
-                {
-                    // Truncate to the max host name length if needed
-                    hostId = hostId.Substring(0, ScriptConstants.MaximumHostIdLength);
-                }
-            }
-
-            // Lowercase and trim any trailing '-' as they can cause problems with queue names
-            return hostId?.ToLowerInvariant().TrimEnd('-');
         }
 
         private void OnUnhandledException(object sender, UnhandledExceptionEventArgs e)
