+++ test/WebJobs.Script.Tests/ScriptHostTests.cs
@@ -1214,38 +1214,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             Assert.False(host.IsFunction(null));
         }
 
-        [Fact]
-        public void GetDefaultHostId_SelfHost_ReturnsExpectedResult()
-        {
-            var config = new ScriptHostConfiguration
-            {
-                IsSelfHost = true,
-                RootScriptPath = @"c:\testing\FUNCTIONS-TEST\test$#"
-            };
-            var scriptSettingsManagerMock = new Mock<ScriptSettingsManager>(MockBehavior.Strict);
-
-            string hostId = ScriptHost.GetDefaultHostId(scriptSettingsManagerMock.Object, config);
-            string sanitizedMachineName = Environment.MachineName
-                    .Where(char.IsLetterOrDigit)
-                    .Aggregate(new StringBuilder(), (b, c) => b.Append(c)).ToString().ToLowerInvariant();
-            Assert.Equal($"{sanitizedMachineName}-789851553", hostId);
-        }
-
-        [Theory]
-        [InlineData("TEST-FUNCTIONS--", "test-functions")]
-        [InlineData("TEST-FUNCTIONS-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", "test-functions-xxxxxxxxxxxxxxxxx")]
-        [InlineData("TEST-FUNCTIONS-XXXXXXXXXXXXXXXX-XXXX", "test-functions-xxxxxxxxxxxxxxxx")] /* 32nd character is a '-' */
-        [InlineData(null, null)]
-        public void GetDefaultHostId_AzureHost_ReturnsExpectedResult(string input, string expected)
-        {
-            var config = new ScriptHostConfiguration();
-            var scriptSettingsManagerMock = new Mock<ScriptSettingsManager>(MockBehavior.Strict);
-            scriptSettingsManagerMock.SetupGet(p => p.AzureWebsiteUniqueSlotName).Returns(() => input);
-
-            string hostId = ScriptHost.GetDefaultHostId(scriptSettingsManagerMock.Object, config);
-            Assert.Equal(expected, hostId);
-        }
-
         public class AssemblyMock : Assembly
         {
             public override object[] GetCustomAttributes(Type attributeType, bool inherit)
